fix: moves testing stuff to appropiatte folder
